+++ crankshaft-console/CHANGELOG.md
@@ -7,6 +7,10 @@ and this project adheres to [Semantic Versioning](https://semver.org/spec/v2.0.0
 
 ## Unreleased
 
+### Added
+
+* Added handling for the task resource usage event ([#86](https://github.com/stjude-rust-labs/crankshaft/pull/86)).
+
 ## 0.1.0 - 08-25-2026
 
 #### Added